Move "storage" classes to data/
