+++ futaba/__main__.py
@@ -23,8 +23,6 @@ from toml import TomlDecodeError
 from . import client
 from .config import InvalidConfigError, load_config
 
-# TODO replace sql "storage" classes with a different name and form
-
 LOG_FILE = "futaba.log"
 LOG_FILE_MODE = "w"
 LOG_FORMAT = "[%(levelname)s] %(asctime)s %(name)s: %(message)s"
