+++ futaba/sql/models/settings.py
@@ -38,118 +38,18 @@ from sqlalchemy import (
 from sqlalchemy.sql import select
 
 from futaba.enums import LocationType
-from futaba.utils import partition_on
+from ..data import (
+    GuildSettingsData,
+    ReapplyRolesData,
+    SpecialRoleData,
+    TrackingBlacklistData,
+)
 from ..hooks import register_hook
 
 Column = functools.partial(Column, nullable=False)
 logger = logging.getLogger(__name__)
 
 __all__ = ["SettingsModel"]
-
-
-class GuildSettingsStorage:
-    __slots__ = ("prefix", "max_delete_messages", "warn_manual_mod_action")
-
-    def __init__(self, prefix, max_delete_messages, *, warn_manual_mod_action):
-        self.prefix = prefix
-        self.max_delete_messages = max_delete_messages
-        self.warn_manual_mod_action = warn_manual_mod_action
-
-
-class SpecialRoleStorage:
-    __slots__ = ("guild", "member_role", "guest_role", "mute_role", "jail_role")
-
-    def __init__(
-        self, guild, member_role_id, guest_role_id, mute_role_id, jail_role_id
-    ):
-        self.guild = guild
-        self.member_role = self._get_role(member_role_id)
-        self.guest_role = self._get_role(guest_role_id)
-        self.mute_role = self._get_role(mute_role_id)
-        self.jail_role = self._get_role(jail_role_id)
-
-    def update(self, attrs):
-        logger.debug("Updating special role storage: %s", attrs)
-        if "member" in attrs:
-            self.member_role = attrs["member"]
-        if "guest" in attrs:
-            self.guest_role = attrs["guest"]
-        if "mute" in attrs:
-            self.mute_role = attrs["mute"]
-        if "jail" in attrs:
-            self.jail_role = attrs["jail"]
-
-    @property
-    def member(self):
-        return self.member_role
-
-    @property
-    def guest(self):
-        return self.guest_role
-
-    @property
-    def mute(self):
-        return self.mute_role
-
-    @property
-    def jail(self):
-        return self.jail_role
-
-    def _get_role(self, id):
-        if id is None:
-            return None
-
-        return discord.utils.get(self.guild.roles, id=id)
-
-    def __iter__(self):
-        yield self.member_role
-        yield self.guest_role
-        yield self.mute_role
-        yield self.jail_role
-
-
-class ReapplyRolesStorage:
-    __slots__ = ("roles", "auto_reapply")
-
-    def __init__(self, roles, auto_reapply):
-        self.roles = roles
-        self.auto_reapply = auto_reapply
-
-
-class TrackingBlacklistStorage:
-    __slots__ = ("guild", "blacklisted_channels", "blacklisted_users")
-
-    def __init__(self, guild, blacklist):
-        # blacklist is an iterable of (type, data_id)
-
-        self.guild = guild
-        blacklisted_channels, blacklisted_users = partition_on(
-            lambda block: block[0] is LocationType.CHANNEL,
-            blacklist,
-            lambda block: block[1],
-        )
-
-        self.blacklisted_channels, self.blacklisted_users = (
-            set(blacklisted_channels),
-            set(blacklisted_users),
-        )
-
-    def is_blocked(self, user_or_channel):
-        if isinstance(user_or_channel, discord.abc.User):
-            return user_or_channel.id in self.blacklisted_users
-        return user_or_channel.id in self.blacklisted_channels
-
-    def add_block(self, user_or_channel):
-        if isinstance(user_or_channel, discord.abc.User):
-            self.blacklisted_users.add(user_or_channel.id)
-        else:
-            self.blacklisted_channels.add(user_or_channel.id)
-
-    def remove_block(self, user_or_channel):
-        if isinstance(user_or_channel, discord.abc.User):
-            self.blacklisted_users.discard(user_or_channel.id)
-        else:
-            self.blacklisted_channels.discard(user_or_channel.id)
 
 
 class SettingsModel:
@@ -266,7 +166,7 @@ class SettingsModel:
             warn_manual_mod_action=False,
         )
         self.sql.execute(ins)
-        self.guild_settings_cache[guild] = GuildSettingsStorage(
+        self.guild_settings_cache[guild] = GuildSettingsData(
             None, self.sql.max_delete_messages, warn_manual_mod_action=False
         )
 
@@ -298,7 +198,7 @@ class SettingsModel:
             self.add_guild_settings(guild)
 
         prefix, max_delete_messages, warn_manual_mod_action = result.fetchone()
-        self.guild_settings_cache[guild] = GuildSettingsStorage(
+        self.guild_settings_cache[guild] = GuildSettingsData(
             prefix, max_delete_messages, warn_manual_mod_action=warn_manual_mod_action
         )
 
@@ -384,9 +284,7 @@ class SettingsModel:
             jail_role_id=None,
         )
         self.sql.execute(ins)
-        self.special_roles_cache[guild] = SpecialRoleStorage(
-            guild, None, None, None, None
-        )
+        self.special_roles_cache[guild] = SpecialRoleData(guild, None, None, None, None)
 
     def remove_special_roles(self, guild):
         logger.info(
@@ -419,7 +317,7 @@ class SettingsModel:
             return self.special_roles_cache[guild]
 
         member_role_id, guest_role_id, mute_role_id, jail_role_id = result.fetchone()
-        roles = SpecialRoleStorage(
+        roles = SpecialRoleData(
             guild, member_role_id, guest_role_id, mute_role_id, jail_role_id
         )
         self.special_roles_cache[guild] = roles
@@ -450,7 +348,7 @@ class SettingsModel:
             guild_id=guild.id, auto_reapply=True, role_ids=[]
         )
         self.sql.execute(ins)
-        self.reapply_roles_cache[guild] = ReapplyRolesStorage(set(), True)
+        self.reapply_roles_cache[guild] = ReapplyRolesData(set(), True)
 
     def remove_reapply_roles(self, guild):
         logger.info(
@@ -482,7 +380,7 @@ class SettingsModel:
             if role is not None:
                 roles.add(role)
 
-        storage = ReapplyRolesStorage(roles, auto_reapply)
+        storage = ReapplyRolesData(roles, auto_reapply)
         self.reapply_roles_cache[guild] = storage
         return storage
 
@@ -585,7 +483,7 @@ class SettingsModel:
         ).where(self.tb_tracking_blacklists.c.guild_id == guild.id)
         result = self.sql.execute(sel)
 
-        blacklist = TrackingBlacklistStorage(guild, result.fetchall())
+        blacklist = TrackingBlacklistData(guild, result.fetchall())
         self.tracking_blacklist_cache[guild] = blacklist
         return blacklist
